Rename to `fetchAllMajorSectors` for consistency
