+++ src/utils/callRustService/lib/fetchAllMajorSectors.ts
